requirements.txt
amqp==2.6.1
asgiref==3.2.7
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.7
decorator==4.4.2
Django==3.0.5
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-mathfilters==1.0.0
entrypoints==0.3
Faker==4.0.3
flake8==3.7.9
gunicorn==20.0.4
importlib-metadata==1.7.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
kombu==4.6.11
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
psycopg2-binary
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redgreenunittest==0.1.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
vine==1.3.0
wcwidth==0.1.9
zipp==3.1.0
django-datetime-widget2
django-widget-tweaks
